--- registry.js
@@ -1,14 +1,0 @@
-/*
-//not viable since it needs explicit uniforms. possible standardization?
-export const SHADER_LIST = [
-	"shader",
-	"shadow"
-]
-*/
-
-export const SOUND_LIST = [
-	//"assets/test.ogg"
-]
-
-export const SPRITE_LIST = [
-]
