+++ state.js
@@ -26,32 +26,13 @@ export let inventory = {
     postits: []
 };
 
-// TODO: Find a better place for this
-export let itemSprites = {
-    0: "assets/Ring_poliert_Blickdicht.png"
-};
-
-export function pickUp(item) {
-	let index = level.objects.indexOf(item);
-	if (index > -1) {
-		level.objects.splice(index, 1);
-        for (let i = 0; i < 40; i++) {
-            let transform = inventoryItemTransform(inventory.objects.length);
-            inventory.objects.push(new Sprite(itemSprites[item.pickup], transform));
-            let m = mat4.create();
-            mat4.fromScaling(m, vec3.fromValues(0.5, 0.5, 1));
-            inventory.postits.push(new Sprite("assets/dull_sticky_bitch.png", mat4.mul(m, transform, m)));
-        }
-	}
-	console.log(inventory);
-}
 
 const INVENTORY_SIZE = 6;
 export const INVENTORY_HEIGHT = 5;
 export const INVENTORY_WIDTH = 9;
 const INVENTORY_SCALE = INVENTORY_SIZE / INVENTORY_HEIGHT;
 
-function inventoryItemTransform(index) {
+export function inventoryItemTransform(index) {
     let x = index % INVENTORY_WIDTH;
     let y = Math.floor(index / INVENTORY_WIDTH);
     let pos = vec3.fromValues((x + 0.5) * INVENTORY_SCALE - INVENTORY_SCALE * INVENTORY_WIDTH / 2, INVENTORY_SIZE / 2 - (y + 0.5) * INVENTORY_SCALE, 0);
